Implement network and compression filtering
